Related to #11 trying to clean up the code so I can get a menu system in place.
