+++ bin/printsly
@@ -13,8 +13,8 @@ puts # formatting
 puts # formatting
 puts "Selectable options will normally be enclosed in " + "[ ]".yellow + "'s."
 puts # formatting
-if File.exists?('../lib/save_game.json')
-  puts "Would you like to start a new game?"
+if File.exists?('../lib/config.json')
+  puts "Would you like to load a config?"
 #  GameSelect.new.outcome
   puts # formatting
 else
